+++ src/modules/login/routes/index.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import { login } from '../controllers/login.controllers';
+import { login } from '../controllers';
 
 const router = Router();
