+++ src/modules/posts/controllers/index.ts
@@ -1,4 +1,4 @@
-import prisma from '../configs/prisma.config';
+import prisma from '../../../configs/prisma.config';
 
 export const getAllPosts: Controller = async (_req, res) => {
   const posts = await prisma.post.findMany();
