+++ src/modules/posts/routes/index.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import { getAllPosts } from '../controllers/posts.controllers';
+import { getAllPosts } from '../controllers';
 
 const router = Router();
